Improve loading of deployment details
